Shifts MilitaryUnit logic into GameObject
Due to an issue with casting gameobjects into MilitaryUnits (the
constructor wasn't initializing properly, due to the style of casting I
was trying to use). While this isn't an ideal solution, with the
gameobject class being muddied a bit, to save time and get something
working this is the best I could think of.